requirements.txt
Flask
azure-identity
numpy
pandas
scikit-learn
matplotlib
nltk
textblob
wordcloud
requests
python-docx
openpyxl
PyPDF2
azure-storage-blob
azure-data-tables
python-dotenv
pyodbc
spacy